ditto/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pyro-ppl>=1.8
torch
